v1/requirements.txt
attrs==24.2.0
Automat==24.8.1
beautifulsoup4==4.12.3
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
constantly==23.10.4
cryptography==43.0.1
cssselect==1.2.0
defusedxml==0.7.1
exceptiongroup==1.2.2
filelock==3.16.1
h11==0.14.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
itemadapter==0.9.0
itemloaders==1.3.1
jmespath==1.0.1
lxml==5.3.0
outcome==1.3.0.post0
packaging==24.1
parsel==1.9.1
pillow==10.4.0
Protego==0.3.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
PyDispatcher==2.0.7
pyOpenSSL==24.2.1
PyPDF2==3.0.1
PySocks==1.7.1
pytesseract==0.3.13
python-dotenv==1.0.1
queuelib==1.7.0
requests==2.32.3
requests-file==2.1.0
Scrapy==2.11.2
scrapy-user-agents==0.1.1
selenium==4.24.0
service-identity==24.1.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
tldextract==5.1.2
tomli==2.0.1
trio==0.26.2
trio-websocket==0.11.1
Twisted==24.7.0
typing_extensions==4.12.2
ua-parser==0.18.0
urllib3==2.2.3
user-agents==2.2.0
w3lib==2.2.1
webdriver-manager==4.0.2
websocket-client==1.8.0
wsproto==1.2.0
zope.interface==7.0.3
